datarush_hackathon/requirements.txt
# requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=1.5.0
google-generativeai>=0.3.0
langchain>=0.0.350
python-dotenv>=1.0.0
numpy>=1.21.0
requests>=2.28.0
openpyxl>=3.0.0
reportlab>=3.6.0
scipy>=1.9.0
seaborn>=0.11.0

# BigQuery Integration (DISABLED - using local data only)
# google-cloud-bigquery>=3.11.0
# google-cloud-bigquery-storage>=2.19.0

# Data Analysis Agent Dependencies (Optional - fallback available)
# google-adk>=0.1.0